release: Volunteer @beorn7 and @csmarchbanks for 2.33 and 2.34
In a meeting at Grafana Labs, we had the shocking revelation that the
next release is scheduled to happen soon, but there is no release
shepherd for it yet.

(Which reminds me: Perhaps we should make it a responsibility of the
release shepherd to make sure there is one for the next release, too?)

Given my current part-time commitment, I would actually try to avoid
being the release shepherd, but on the other hand, it's about time
that a Grafanista to do the job again, and the only one who could do
it at all in the room was me. @csmarchbanks volunteered for the next
release. So here we are.

Of course, if any other team member would like to jump in and do this
release or the next, we'll happily insert you into the list.

Signed-off-by: beorn7 <beorn@grafana.com>