Use sync.Mutex for chunk write queue locks
sync.RWMutex seem to be starving get operations when there's a lot of write operations because AFAIK it's a write preferring lock.
During compaction or other operation that end up issuing a lot of chunk writes this lock can starve reads and cause a visible delay for rule evaluation.
Switch to sync.Mutex so that all operations are equal and reads have a better chance of being serviced in time.

Fixes #10377

Signed-off-by: Łukasz Mierzwa <l.mierzwa@gmail.com>