POC: Load service discoveries from plugins (*.so files)
This is a proof-of-concept to move the Azure SD to a signed shared file.
Azure is just an example, the goal is not to move it out-of-tree, rather
to experiment with it in the scope of this pull request.

The goal is to be able to release service discoveries as part of the
Prometheus release process without incrasing the binary size.

Plugins can be generated like this:

cd discovery/azure
go build -buildmode=plugin

To make sure that no third party code is loaded into prometheus, the so
files have to be signed:

cosign generate-key-pair
cp cosign.pub cmd/prometheus
cosign --key cosign.key sign-blob discovery/azure/azure.so > discovery/azure/azure.so.cosign

Prometheus can be launched with this:

./prometheus --enable-feature=plugins --plugin discovery/azure/azure.so

There are a few things to note:
- We could generate temporary keys on each build, and the users would
  have to get a new so file at each prometheus release.
  This is very likely to be the case anyway, because imports must all be
  of the same version.
- Currently the azure.so file is big, but moving the discoverer registry to
  another package could help here. It might also be because it is
  imported in the config package tests.
- Windows does not support this, so this would only be for new service
  discoveries. That would remove feature parity between OS'es.

Signed-off-by: Julien Pivotto <roidelapluie@inuits.eu>