Proper support for go modules.
This pull requests makes Prometheus go-mod compatible.

The general idea is to release the Prometheus libraries as v0.x
releases, and the Prometheus binaries as v2.x.

This is done my moving the cmd package to a v2 directory.
Then, we would tag v0.35.0-beta.0, which would make the libraries v0.35.0-beta.0
available for consumers.

When v2.X.0 is released, we would release v0.X.0.

Pre-go mod versions are retracted from go.mod. This is not nice but
should work.

Only v2.x tags will be built and released by CI. v0.x.x tags would just
be normal tags in the repo, not promoted as releases.

Additional benefits:
- Prometheus is installable with go install
  github.com/prometheus/prometheus/v2/cmd/prometheus@v2.35.0
- Does not require go 1.18 workspace magic.

Signed-off-by: Julien Pivotto <roidelapluie@inuits.eu>