Abort WAL replay if Prometheus is stopping
WAL replay can take a long time depending on the amount of scraped metrics. In our environment it often takes 20-40 minutes to complete it with 20-30M time series.
At the same time this process cannot be interrupted, which means that if I start Prometheus but decide I need to restart it for some reason before it completes WAL replay, I will have to wait twice for the WAL replay to finish - that's because it will only shutdown after it completes WAL replay and then when it starts again the whole process starts from scratch.
This change passes context down to WAL replay code so we can cancel it at any point, which will aboirt WAL replay and allow Prometheus to skip the long wait just to restart the whole process again.

Signed-off-by: Łukasz Mierzwa <l.mierzwa@gmail.com>