LTS: Implement Go version upgrade and switch to GitHub Actions
This commit upgrades the underlying Go version from 1.18 to 1.19 in
order to align with the Prometheus main branch and ensure that the
Long-Term Support (LTS) release remains under a supported version of Go.
The upgrade also requires changes to the build process for Windows.
Additionally, this commit switches the CI/CD pipeline from CircleCI to
GitHub Actions for better integration and alignment with the Prometheus
main branch.

This commit also runs go fmt on the entire codebase using Go 1.19.

This is following an email sent to the prometheus-developers mailing
list in december.

Signed-off-by: Julien Pivotto <roidelapluie@o11y.eu>