tsdb agent: drop deleted series from WAL sooner
`head.deleted` holds the WAL segment in use at the time each series was
removed from the head. At the end of `truncateWAL()` we will delete
all segments up to `last`, so we can drop any series that were last seen
in a segment at or before that point.

Same change as in `tsdb/head.go`.

Signed-off-by: Bryan Boreham <bjboreham@gmail.com>