mv labels_string.go labels_as_string.go
This is a minor cosmetical change, but my IDE (and I guess many of them)
nests `labels_string.go` under `labels.go` because it assumes it's the
file generated by the `stringer` tool, which follows that naming
pattern.

We can blame my IDE, but it's also not a big deal to rename it, right?

Signed-off-by: Oleg Zaytsev <mail@olegzaytsev.com>