Additionally wrap WBL replay error
Although WBL replay is already wrapped with errLoadWbl,
there are other errors that can happen during a WBL replay.
We should not try to repair WAL in those cases.

This commit additionally wraps the final error in Head.Init again
with errLoadWbl so that WBL replay errors can be identified properly.

Since we double wrap it, we unwrap it before the wbl repair.

Alternative solution is to return a boolean from Head.Init to identify
the error source, but that looks less clean and leads to a LOT of refactor.

Signed-off-by: Ganesh Vernekar <ganeshvern@gmail.com>