head_append.go: use new AppendOrCreate histogram functions
AppendOrCreate histogram functions enhanced with optional previous
appender parameter to let the appender know if it following a previous
chunk or not. This is needed when a new chunk is cut outside the appender,
for example because of size limit.

This now fully hides handling the counter reset header. This means the
chunk specific SetCounterResetHeader is made private.

Signed-off-by: György Krajcsovits <gyorgy.krajcsovits@grafana.com>