scraping: limit dropped targets, to save memory
It's possible (quite common on Kubernetes) to have a service discovery
return thousands of targets then drop most of them in relabel rules.
The main place this data is used is to display in the web UI, where
you don't want thousands of lines of display.

Set a default limit of 100 per job.

Signed-off-by: Bryan Boreham <bjboreham@gmail.com>